Separated GameObjectExtensions class into a partial multi file class
