+++ Scripts/ReplaceChild.cs.meta
@@ -1,5 +1,7 @@
 fileFormatVersion: 2
-guid: 44a4cf46cc0e349c9a0b710803bafb9f
+guid: 6377654a063924155a46ab557c9db197
+timeCreated: 1528118410
+licenseType: Pro
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
